+++ src/components/admin/AdminHeader.css
@@ -1,0 +1,77 @@
+/* Admin Header Styles */
+.admin-header {
+  background-color: #1E1E1E;
+  padding: 1rem 1.5rem;
+}
+
+.admin-header-container {
+  max-width: 1200px;
+  margin: 0 auto;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.admin-header-left {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.admin-logo {
+  display: flex;
+  align-items: center;
+}
+
+.admin-logo-image {
+  height: 38px;
+  width: auto;
+  border-radius: 50%;
+}
+
+.admin-header-title {
+  display: flex;
+  flex-direction: column;
+}
+
+.admin-page-type {
+  color: var(--color-white);
+  font-size: 1.3rem;
+  font-weight: 500;
+  line-height: 1.2;
+}
+
+.admin-page-title {
+  color: var(--color-white);
+  font-size: 0.875rem;
+  font-weight: 500;
+  line-height: 1.0;
+}
+
+.admin-header-right {
+  display: flex;
+  align-items: center;
+}
+
+.logout-btn {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  background: none;
+  border: none;
+  color: var(--color-white);
+  cursor: pointer;
+  padding: 0.5rem 1rem;
+  border-radius: 4px;
+  font-size: 0.875rem;
+  font-weight: 500;
+  transition: background-color 0.2s ease;
+}
+
+.logout-btn:hover {
+  background-color: rgba(255, 255, 255, 0.1);
+}
+
+.logout-btn:active {
+  background-color: rgba(255, 255, 255, 0.2);
+}
