+++ src/components/admin/index.js
@@ -1,0 +1,1 @@
+export { default as AdminHeader } from './AdminHeader';
