+++ src/components/common/LazyComponents.jsx
@@ -2,7 +2,7 @@ import { lazy } from 'react';
 
 // 페이지 컴포넌트들 지연 로딩
 export const LazyMainPage = lazy(() => import('../../pages/MainPage'));
-export const LazyAdminLoginPage = lazy(() => import('../../pages/AdminLoginPage'));
+export const LazyAdminLoginPage = lazy(() => import('../pages/AdminLogin/AdminLoginPage'));
 export const LazyRecruitingManagePage = lazy(() => import('../../pages/RecruitingManagePage'));
 export const LazyRecruitingDetailPage = lazy(() => import('../../pages/RecruitingDetailPage'));
