--- src/components/main/CategoriesSection.css
@@ -1,108 +1,0 @@
-/* 모집 대상 섹션 */
-.main-page-categories {
-  padding: 6rem 0;
-  position: relative;
-  overflow: hidden;
-}
-
-.main-page-categories::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none;
-}
-
-.main-page-categories-content {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 0 2rem;
-  text-align: center;
-  position: relative;
-  z-index: 1;
-}
-
-.main-page-categories-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 2rem;
-  margin: 3rem 0;
-  max-width: 800px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.main-page-category-card {
-  background: rgba(255, 255, 255, 0.05);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: 20px;
-  padding: 2rem 1.5rem;
-  text-align: center;
-  transition: all 0.3s ease;
-  backdrop-filter: blur(10px);
-}
-
-.main-page-category-card:hover {
-  transform: translateY(-10px);
-  background: rgba(255, 255, 255, 0.08);
-  border-color: var(--color-green-0);
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.3);
-}
-
-.main-page-category-icon {
-  font-size: 3rem;
-  margin-bottom: 1rem;
-  display: block;
-}
-
-.main-page-category-title {
-  color: var(--color-white);
-  font-size: 1.5rem;
-  font-weight: 600;
-  margin: 0;
-}
-
-.main-page-categories-description {
-  color: var(--color-gray-1);
-  font-size: 1.125rem;
-  line-height: 1.6;
-  max-width: 600px;
-  margin: 3rem auto 0;
-}
-
-
-/* 반응형 디자인 */
-@media (max-width: 768px) {
-  .main-page-categories {
-    padding: 4rem 0;
-  }
-  
-  .main-page-categories-content {
-    padding: 0 1rem;
-  }
-  
-  .main-page-categories-grid {
-    grid-template-columns: 1fr;
-    gap: 1.5rem;
-    margin: 2rem 0;
-  }
-  
-  .main-page-category-card {
-    padding: 1.5rem 1rem;
-  }
-  
-  .main-page-category-icon {
-    font-size: 2.5rem;
-  }
-  
-  .main-page-category-title {
-    font-size: 1.25rem;
-  }
-  
-  .main-page-categories-description {
-    font-size: 1rem;
-    margin-top: 2rem;
-  }
-}
