--- src/components/main/InfoSection.css
@@ -1,46 +1,0 @@
-/* FAQ & 모집 일정 섹션 */
-.main-page-info-section {
-  padding: 5rem 0;
-  text-align: center;
-  position: relative;
-}
-
-.main-page-info-section::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none;
-}
-
-.main-page-info-content {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 0 2rem;
-  position: relative;
-  z-index: 1;
-}
-
-.main-page-info-message {
-  color: var(--color-gray-1);
-  font-size: 1.125rem;
-  line-height: 1.6;
-  margin-top: 1.5rem;
-}
-
-/* 반응형 디자인 */
-@media (max-width: 768px) {
-  .main-page-info-section {
-    padding: 3rem 0;
-  }
-  
-  .main-page-info-content {
-    padding: 0 1rem;
-  }
-  
-  .main-page-info-message {
-    font-size: 1rem;
-  }
-}
