+++ src/components/pages/AdminLogin/components/LoginForm.css
@@ -1,23 +1,4 @@
-/* Admin Login Page Styles */
-.admin-login-page-unique {
-  min-height: 100vh;
-  background-color: var(--color-black);
-  font-family: var(--font-family);
-}
-
-.admin-login-main {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: calc(100vh - 80px);
-  padding: 2rem;
-}
-
-.admin-login-container {
-  width: 100%;
-  max-width: 400px;
-}
-
+/* Login Form Specific Styles */
 .admin-login-form-wrapper {
   background-color: var(--color-black);
   padding: 2rem;
@@ -96,10 +77,6 @@
 
 /* Responsive */
 @media (max-width: 480px) {
-  .admin-login-main {
-    padding: 1rem;
-  }
-  
   .admin-login-form-wrapper {
     padding: 1.5rem;
   }
