+++ src/components/pages/AdminLogin/components/index.js
@@ -1,0 +1,1 @@
+export { default as LoginForm } from './LoginForm';
