+++ src/components/pages/AdminLogin/index.js
@@ -1,0 +1,1 @@
+export { default as AdminLoginPage } from './AdminLoginPage';
