--- src/components/sections/InfoSection && cp CUsersrlarbcodingrecruitpiro-recruit-FEsrccomponentsmainInfoSection.css CUsersrlarbcodingrecruitpiro-recruit-FEsrccomponentssectionsInfoSection
@@ -1,16 +1,0 @@
-import React from 'react';
-import '../../styles/common-sections.css';
-import './InfoSection.css';
-
-const InfoSection = () => {
-  return (
-    <section className="main-page-info-section">
-      <div className="main-page-info-content">
-        <h2 className="main-page-section-title">FAQ & 모집 일정</h2>
-        <p className="main-page-info-message">자세한 질문과 답변 및 모집일정은 메인 페이지를 참고해주세요</p>
-      </div>
-    </section>
-  );
-};
-
-export default InfoSection;
